перенос: сбор вариантов — шесть разделов, площадка и вес замысла
actions.py до строки 1758: мерки дня, корзина, мена и четыре раздела
сбора. С ними meeting.оценка и вес замысла — их зовёт раздел «себя».

Две ловушки, найденные этим переносом и дописанные в ПОРТ.md:
sum() в Python считает компенсированной суммой, а не слева направо
(8 расхождений из 8); округление до шестого знака само создавало
расхождение на ровной половине — убрано из снимка, заменено допуском
в сверке. Сверка от этого стала строже: подложенные 1e-7 краснеют.

+++ Дом.Ядро/Действия.Корзина.cs
@@ -67,7 +67,7 @@ public sealed class Корзина
 
     /// <summary>Сколько раз сегодня уже выходил к соседям и сколько ещё можно.</summary>
     public int визитов { get; }
-    public int предел_визитов { get; }
+    public double предел_визитов { get; }
 
     public List<((string key, object? target) что, double оценка)> варианты { get; } = new();
 
@@ -79,7 +79,8 @@ public sealed class Корзина
     public List<(string key, object? target, string почему)> отказы { get; } = new();
     public Dictionary<КлючЦели, object?> детали { get; } = new();
 
-    public Корзина(House h, NPC npc, int визитов, int предел_визитов, Самочувствие? с = null)
+    public Корзина(House h, NPC npc, int визитов, double предел_визитов,
+                   Самочувствие? с = null)
     {
         this.h = h;
         this.npc = npc;
